blog fix
