+++ src/pages/DefaultRoutes/blog/index.css
@@ -1,3 +1,9 @@
+.swiper-pagination-bullet-active {
+  width: 16px !important;
+  border-radius: 25% !important;
+  transition-duration: 500ms;
+}
+
 .blog-card {
   background-color: #fff;
 }
@@ -8,15 +14,9 @@
 }
 
 .blurry-text {
+  height: 14px;
   color: transparent !important;
-  text-shadow: 3.5px #292a2b;
-}
-
-.blurry-text-md div,
-.blurry-text-md div h1,
-.blurry-text-md article {
-  color: transparent !important;
-  text-shadow: 10px #58595a;
+  background-color: #dadadb;
 }
 
 .blog-article {
@@ -28,6 +28,10 @@
   margin: 12px 0 !important;
 }
 
+.blog-article .blurry-text {
+  margin: 3px 0 !important;
+}
+
 .blog-article h1,
 .blog-article h2,
 .blog-article h3,
@@ -37,6 +41,7 @@
 
 .blog-article p,
 .blog-article span {
+  font-size: 14px !important;
   font-family: "Avenir", sans-serif !important;
 }
